Development of Moderate Apps
I adjusted the User Database so that we have the three roles required before.

Non-users have the ability to access the store, but can't submit app requests

Users can submit applications

Moderators and Admins can now accept or reject before stuff is posted to the app store.

Most of this work is stored in SubmitApp.aspx, .cs,

This took way longer than expected.

~Noah